requirements.txt
Django==1.6.4
Pillow==2.4.0
django-appconf==0.6
django-mptt==0.6.0
django-staticfiles==1.2.1
django-suit==0.2.8
django-tinymce==1.5.2
six==1.6.1
